15-01-2019

+++ src/main/java/br/com/original/rf2bpm/repository/ArquivoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.com.original.rf2bpm.entity.Arquivo;
 
 
-public interface ArquivoRepository extends JpaRepository<Arquivo, Integer> {
+public interface ArquivoRepository extends JpaRepository<Arquivo, Long> {
 
 }
